requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
colorlog==5.0.1
concurrent-log-handler==0.9.9
diff-match-patch==20200713
Django==2.2.2
django-crispy-forms==1.8.0
django-filter==2.4.0
django-formtools==2.1
django-import-export==1.2.0
django-reversion==2.0.0
djangorestframework==3.12.4
elasticsearch==7.12.1
future==0.15.2
har2case==0.3.1
httplib2==0.9.2
HttpRunner==1.5.8
idna==2.8
importlib-metadata==4.0.1
iniconfig==1.1.1
Jinja2==3.0.1
kafka-python==1.4.6
lxml==4.2.5
MarkupSafe==2.0.1
more-itertools==8.7.0
mysqlclient==2.0.3
nb-log==3.4
packaging==20.9
pika==1.2.0
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pymongo==3.5.1
PyMySQL==0.9.3
pyparsing==2.4.7
pypiwin32==223
pytest==4.5.0
pytest-assume==2.4.2
pytest-base-url==1.4.2
pytest-html==1.19.0
pytest-metadata==1.11.0
pytz==2021.1
pywin32==300
PyYAML==5.3.1
requests==2.22.0
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.10.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
tomorrow3==1.1.0
typing-extensions==3.10.0.0
urllib3==1.25.11
wcwidth==0.2.5
zipp==3.4.1
